requirements.txt
nltk==3.4.1
spacy==2.1.3
PyPDF2==1.26.0
fuzzywuzzy==0.17.0
pdfminer.six==20181108
scikit_learn==0.21.1
